# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: posctl API
  version: 1.0.0
  description: |
    Source-of-truth contract for the SantimPay POS Management Platform. The Spring Boot backend
    publishes a runtime spec at /v3/api-docs; THIS file is the hand-curated contract used to generate
    the TypeScript (web) and Dart (field app) clients. A CI contract test reconciles the two.

    Covers all 12 modules. Action endpoints use the `{id}:verb` style (e.g. /merchants/{id}:activate)
    matching the controllers. All endpoints require a Keycloak bearer token; fine-grained
    authorization is enforced server-side against the `permissions` claim.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /merchants:
    get:
      operationId: listMerchants
      tags: [Merchants]
      parameters:
        - { $ref: '#/components/parameters/Q' }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/MerchantStatus' } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { description: page of merchants, content: { application/json: { schema: { $ref: '#/components/schemas/MerchantPage' } } } } }
    post:
      operationId: onboardMerchant
      tags: [Merchants]
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OnboardMerchantRequest' } } }
      responses:
        '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/Merchant' } } } }
        '422': { $ref: '#/components/responses/Problem' }
  /merchants/{id}:
    get:
      operationId: getMerchant
      tags: [Merchants]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: merchant, content: { application/json: { schema: { $ref: '#/components/schemas/Merchant' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /merchants/{id}:activate:
    post:
      operationId: activateMerchant
      tags: [Merchants]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: activated, content: { application/json: { schema: { $ref: '#/components/schemas/Merchant' } } } }
        '409': { $ref: '#/components/responses/Problem' }
  /merchants/{id}:suspend:
    post:
      operationId: suspendMerchant
      tags: [Merchants]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { reason: { type: string } }, required: [reason] } } }
      responses: { '200': { description: suspended, content: { application/json: { schema: { $ref: '#/components/schemas/Merchant' } } } } }

  # ---------------- Devices ----------------
...
  /deployments:
    get:
      operationId: listDeployments
      tags: [Deployments]
      parameters:
        - { in: query, name: date, schema: { type: string, format: date } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/DeploymentStatus' } }
        - { in: query, name: agent, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { description: page of deployments, content: { application/json: { schema: { $ref: '#/components/schemas/DeploymentPage' } } } } }
    post:
      operationId: planDeployment
      tags: [Deployments]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlanDeploymentRequest' } } }
      responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/Deployment' } } } } }
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [Deployments]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: deployment, content: { application/json: { schema: { $ref: '#/components/schemas/Deployment' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /deployments/{id}:complete:
    post:
      operationId: completeDeployment
      tags: [Deployments]
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CompleteDeploymentRequest' } } }
      responses:
        '200': { description: completed, content: { application/json: { schema: { $ref: '#/components/schemas/Deployment' } } } }
        '409': { $ref: '#/components/responses/Problem' }

  # ---------------- KYC ----------------
...
  /tasks:
    get:
      operationId: listTasks
      tags: [Tasks]
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/TaskStatus' } }
        - { in: query, name: assignee, schema: { type: string, format: uuid } }
        - { in: query, name: type, schema: { type: string } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { description: page of tasks, content: { application/json: { schema: { $ref: '#/components/schemas/TaskPage' } } } } }
    post:
      operationId: createTask
      tags: [Tasks]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateTaskRequest' } } }
      responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } } }
  /tasks/{id}:assign:
    post:
      operationId: assignTask
      tags: [Tasks]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { assigneeId: { type: string, format: uuid } }, required: [assigneeId] } } }
      responses: { '200': { description: assigned, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } } }
  /tasks/{id}:complete:
    post:
      operationId: completeTask
      tags: [Tasks]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses: { '200': { description: completed, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } } }

  # ---------------- Follow-ups ----------------
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    Q: { in: query, name: q, schema: { type: string } }
    Page: { in: query, name: page, schema: { type: integer, default: 0 } }
    Limit: { in: query, name: limit, schema: { type: integer, default: 50, maximum: 200 } }
    IdempotencyKey: { in: header, name: Idempotency-Key, schema: { type: string } }
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    PageMeta:
      type: object
      properties:
        limit: { type: integer }
        number: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }

    # Identity
...
    MerchantStatus: { type: string, enum: [ONBOARDING, PENDING_KYC, ACTIVE, SUSPENDED, CLOSED] }
    OnboardMerchantRequest:
      type: object
      required: [merchantNo, legalName]
      properties:
        merchantNo: { type: string, maxLength: 40 }
        legalName: { type: string, maxLength: 200 }
        tradeName: { type: string }
        taxId: { type: string }
        category: { type: string }
    Merchant:
      type: object
      properties:
        id: { type: string, format: uuid }
        merchantNo: { type: string }
        legalName: { type: string }
        tradeName: { type: string }
        taxId: { type: string }
        category: { type: string }
        status: { $ref: '#/components/schemas/MerchantStatus' }
        riskTier: { type: string, enum: [LOW, MEDIUM, HIGH] }
        onboardedAt: { type: string, format: date-time }
        activatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer }
    MerchantPage:
      type: object
      properties: { data: { type: array, items: { $ref: '#/components/schemas/Merchant' } }, page: { $ref: '#/components/schemas/PageMeta' } }

    # Device
...
    DeploymentStatus: { type: string, enum: [PLANNED, IN_PROGRESS, COMPLETED, FAILED, CANCELLED] }
    PlanDeploymentRequest:
      type: object
      required: [scheduledDate, merchantId, branchId]
      properties:
        deploymentNo: { type: string }
        scheduledDate: { type: string, format: date }
        merchantId: { type: string, format: uuid }
        branchId: { type: string, format: uuid }
        assignedAgent: { type: string, format: uuid }
    CompleteDeploymentRequest:
      type: object
      required: [deviceId]
      properties:
        deviceId: { type: string, format: uuid }
        receivedBy: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        conversationNotes: { type: string }
        trelloCardId: { type: string }
    Deployment:
      type: object
      properties:
        id: { type: string, format: uuid }
        deploymentNo: { type: string }
        scheduledDate: { type: string, format: date }
        merchantId: { type: string, format: uuid }
        branchId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        assignedAgent: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/DeploymentStatus' }
        receivedBy: { type: string }
        completedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer }
    DeploymentPage:
      type: object
      properties: { data: { type: array, items: { $ref: '#/components/schemas/Deployment' } }, page: { $ref: '#/components/schemas/PageMeta' } }

    # KYC
...
    TaskStatus: { type: string, enum: [OPEN, ASSIGNED, IN_PROGRESS, BLOCKED, DONE, CANCELLED] }
    TaskPriority: { type: string, enum: [LOW, MEDIUM, HIGH, URGENT] }
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 200 }
        description: { type: string }
        taskType: { type: string }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        relatedType: { type: string }
        relatedId: { type: string, format: uuid }
        dueAt: { type: string, format: date-time }
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        taskType: { type: string }
        priority: { $ref: '#/components/schemas/TaskPriority' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        assigneeId: { type: string, format: uuid }
        relatedType: { type: string }
        relatedId: { type: string, format: uuid }
        dueAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        source: { type: string }
        createdAt: { type: string, format: date-time }
    TaskPage:
      type: object
      properties: { data: { type: array, items: { $ref: '#/components/schemas/Task' } }, page: { $ref: '#/components/schemas/PageMeta' } }

    # Follow-up
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }